+++ lib/arachni/element/cookie.rb
@@ -175,7 +175,7 @@ class Cookie < Base
     end
 
     # Overrides {Capabilities::Mutable#each_mutation} to handle cookie-specific
-    # limitations and the {Arachni::Options#audit_cookies_extensively} option.
+    # limitations and the {Arachni::OptionGroups::Audit#cookies_extensively} option.
     #
     # @param (see Capabilities::Mutable#each_mutation)
     # @return (see Capabilities::Mutable#each_mutation)
@@ -184,12 +184,15 @@ class Cookie < Base
     #
     # @see Capabilities::Mutable#each_mutation
     def each_mutation( payload, opts = {}, &block )
-        flip = opts.delete( :param_flip )
+        opts        = opts.dup
+        flip        = opts.delete( :param_flip )
+        extensively = opts[:extensively]
+        extensively = Arachni::Options.audit.cookies_extensively? if extensively.nil?
 
         super( payload, opts ) do |elem|
             yield elem
 
-            next if !Arachni::Options.audit.cookies_extensively?
+            next if !extensively
             elem.each_extensive_mutation( elem, &block )
         end